Added custom generator sorting and analytics
- Contributed by Adam Burgess

+++ Dialogs/Generator/Top10StrategyInfo.cs
@@ -4,11 +4,13 @@
 // Copyright (c) 2006 - 2012 Miroslav Popov - All rights reserved.
 // This code or any part of it cannot be used in other applications without a permission.
 
+using System;
+
 namespace Forex_Strategy_Builder.Dialogs.Generator
 {
     public class Top10StrategyInfo
     {
-        public int Balance { get; set; }
+        public Single Balance { get; set; }
         public Top10Slot Top10Slot { get; set; }
         public Strategy TheStrategy { get; set; }
     }
